// frontend/src/lib/api/models/DetailedSummaryContent.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { SummaryDefinition } from './SummaryDefinition';
export type DetailedSummaryContent = {
    overview: string;
    keyConcepts: Array<string>;
    importantDefinitions: Array<SummaryDefinition>;
    mainExplanations: Array<string>;
    examples: Array<string>;
    examRelevantPoints: Array<string>;
    labNotes?: (Array<string> | null);
};